smartquizjr_backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
sqlmodel
pydantic
psycopg2-binary
sqlalchemy